Refactor UserInterface class to handle score display and moved helper methods to helpers folder
